wip: add room support

+++ src/player/game_data.rs
@@ -3,6 +3,6 @@ use serde::Serialize;
 #[derive(Default, Serialize)]
 pub struct GameData {
     pub sprite: String,
-    pub sprite_index: isize,
+    pub sprite_index: u32,
     pub system: String,
 }
